% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{chinntoda}
\alias{chinntoda}
\title{Translate a single Chinse Date or Chinese number into Arabic number}
\usage{
chinntoda(input)
}
\arguments{
\item{input}{input chinese sentences}
}
\value{
numeric
}
\description{
Translate a single Chinse Date or Chinese number into Arabic number
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

