% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{cnextract}
\alias{cnextract}
\title{Cut a chinese sentence based on a given start and end}
\usage{
cnextract(input, start, end)
}
\arguments{
\item{input}{input chinese sentences}

\item{start}{starting words}

\item{end}{ending words}
}
\value{
list
}
\description{
Cut a chinese sentence based on a given start and end
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

