% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{codemoney}
\alias{codemoney}
\title{Translate Chinese number into Arabic number}
\usage{
codemoney(input)
}
\arguments{
\item{input}{formatted Chinese numer}
}
\value{
numeric
}
\description{
Translate Chinese number into Arabic number
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

