% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{codemoneyv}
\alias{codemoneyv}
\title{Translate Chinese number into Arabic number vectorize verison
Use recursive method}
\usage{
codemoneyv(input)
}
\arguments{
\item{input}{formatted Chinese numer}
}
\value{
numeric
}
\description{
Translate Chinese number into Arabic number vectorize verison
Use recursive method
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

