% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{codevar}
\alias{codevar}
\title{Code the legal factor.the logic is extract sentences that has 
certain words and it is not negative sentences}
\usage{
codevar(input, findit, sep = c(",", "."), detectneg = TRUE)
}
\arguments{
\item{input}{formatted Chinese numer}

\item{findit}{sentence seperator}

\item{sep}{sentence seperator sep= comma dot}

\item{detectneg}{whether to turn on detect negative function}
}
\description{
Code the legal factor.the logic is extract sentences that has 
certain words and it is not negative sentences
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

