% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{findpos}
\alias{findpos}
\title{Find the position of a certain Chinese word.}
\usage{
findpos(input, find)
}
\value{
Ture/False
}
\description{
Find the position of a certain Chinese word.
}
\examples{

}
\keyword{basic}

