% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{is.num_coma}
\alias{is.num_coma}
\title{Detech whether a given chinese word is numeric or "."}
\usage{
is.num_coma(input)
}
\arguments{
\item{input}{input chinese sentences}
}
\value{
TRUE or FALSE
}
\description{
Detech whether a given chinese word is numeric or "."
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{basic}

