% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{keepselect}
\alias{keepselect}
\title{Cut sentence based on given characters vector A and keep the sentences contains vector B}
\usage{
keepselect(input, findit, sep = c(",", "."))
}
\arguments{
\item{input}{formatted Chinese numer}

\item{findit}{sentence seperator}

\item{sep}{sentence seperator sep= comma, dot}
}
\description{
Cut sentence based on given characters vector A and keep the sentences contains vector B
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

