% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{cutsentence}
\alias{cutsentence}
\title{Cut the Chinese sentences by given characters.}
\usage{
cutsentence(input, find)
}
\arguments{
\item{input}{input chinese sentences}

\item{find}{cut by which}
}
\value{
as list
}
\description{
Cut the Chinese sentences by given characters.
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{basic}

