% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{detectnegative}
\alias{detectnegative}
\title{Detect whether there is negative adverb in Chinese sentence
不、非、无、未、不曾、没、没有、请勿、不用、无须、并非、毫无、决不、休想、
永不、不要、未尝、未曾、毋、莫、从不、从未、从未有过、尚未、一无、并未、尚无、
从来不、从没、绝非、远非、切莫、永不、休想、绝不、毫不、不必、禁止、忌、拒绝、
杜绝、否、弗、木有}
\usage{
detectnegative(input, findit, sep = c("<U+00A3><U+00AC>", "<U+00A1><U+00A3>"))

detectnegative(input, findit, sep = c("<U+00A3><U+00AC>", "<U+00A1><U+00A3>"))
}
\arguments{
\item{input}{formatted Chinese numer}

\item{findit}{sentence seperator}

\item{sep}{sentence seperator}

\item{input}{formatted Chinese numer}
}
\value{
numeric

numeric
}
\description{
Detect whether there is negative adverb in Chinese sentence
不、非、无、未、不曾、没、没有、请勿、不用、无须、并非、毫无、决不、休想、
永不、不要、未尝、未曾、毋、莫、从不、从未、从未有过、尚未、一无、并未、尚无、
从来不、从没、绝非、远非、切莫、永不、休想、绝不、毫不、不必、禁止、忌、拒绝、
杜绝、否、弗、木有

Cut sentence based on given characters vector A and keep the sentences contains vector B
}
\examples{


}
\author{
Xia Yiwei

Xia Yiwei
}
\keyword{ADV}

