% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{detectnegative}
\alias{detectnegative}
\title{Detect whether there is negative adverb in Chinese sentence}
\usage{
detectnegative(input)
}
\arguments{
\item{input}{formatted Chinese numer}
}
\value{
numeric
}
\description{
Detect whether there is negative adverb in Chinese sentence
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{ADV}

