% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{findpos}
\alias{findpos}
\title{Find the position of a certain Chinese word.}
\usage{
findpos(input, find)
}
\arguments{
\item{input}{input chinese sentences}

\item{find}{find with Chinese words}
}
\value{
Reture number indicate the position of Chinese word, NA if not find
}
\description{
Find the position of a certain Chinese word.
}
\examples{

}
\author{
Xia Yiwei
}
\keyword{basic}

