% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legalwordproc_cn.R
\name{ischinexist}
\alias{ischinexist}
\title{Check whether a certain word is exist.}
\usage{
ischinexist(input, find)
}
\value{
Ture/False
}
\description{
Check whether a certain word is exist.
}
\examples{

}
\keyword{basic}

